+++ dpp/src/document/state_transition/documents_batch_transition/validation/state/validate_documents_batch_transition_state.rs
@@ -1,5 +1,5 @@
 use super::{
-    execute_data_triggers, fetch_documents_factory::fetch_documents,
+    execute_data_triggers::execute_data_triggers, fetch_documents_factory::fetch_documents,
     validate_documents_uniqueness_by_indices::validate_documents_uniqueness_by_indices,
 };
 use crate::{
